+++ app/Http/Controllers/SchoolController.php
@@ -41,8 +41,12 @@ class SchoolController extends Controller
     }
 
     public function destroy(School $school) {
+        if ($school->batchSchool()->count() > 0) {
+            return redirect()->route('school.index')->with('error', 'Tidak dapat menghapus data ini karena memiliki relasi dengan data yang lain.');
+        }
+        
         $school->delete();
 
-        return redirect()->route('school.index');
+        return redirect()->route('school.index')->with('success', 'Data berhasil dihapus.');
     }
 }
